Move documentation to the root
Gerrit libModules are not loaded as Gerrit plugins
and their documentation is not visible from the UI. Therefore
there is no point in including their documentation in the
src/main/resource/Documentation folder as it won't be accessible
anyway.

Also remote the @PLUGIN@ placeholders that won't be expanded
at runtime.

Change-Id: I9b155a2df221a90fff59262fe6b1f7bd10992d9d