Update patch set 1
Patch Set 1:

Let me answer the one by one, at my best knowledge, which is limited.

> Patch Set 1:
> 
> I dived deep into gerrit cache internals I am a bit confused about some things.
> I could do with some guidance.
>  
> There are 3 configuration settings that are relevant to this change:
> 
> maxAge [1], expireAfterWrite [2], refreshAfterWrite[3]
> 
> * "maxAge" it is used to populate the expireAfterWrite setting[4].
> * "expireAfterWrite" is never read from the configuration.
> * "refreshAfterWrite"
> 
> H2CacheImpl looks at expireAfterWrite in two occasions:
> - When calling getIfPresent (from memory), and invalidating the result, if the entry expired.
> - When calling executing scheduled prune(). This is scheduled hardcoded every 30 seconds[5].
> 
> My questions are:
> - is it expected that expireAfterWrite configuration is never read and that maxAge effectively replaces it. What's the difference between the two?

Yes, they are the same concept.

> - Is it expected that H2 database entries are pruned every 30 seconds? if maxAge was set to 10 minutes, entries could potentially survive for 10 minutes and 30 seconds.

The 30s polling is an implementation detail. Also the pruning is not supposed to be done in "real-time" but "some time after the expiry". We are talking about caches and not transactional or real-time systems.

> - For chronicle-map it is unfeasible to scan through every entry in the store to check if they have expired. H2 can leverage ordering to short-circuit[6] but we don't have the same possibility with chronicle-map.

We do need, however, a cleanup mechanism, otherwise the persistent cache would grow indefinitely.

> How about expiring the value *when* they get accessed, rather than scheduling a pruning that would cause a scan of the entire entry set? Ideas how to do this in a timely manner?

That is more a "refreshAfterWrite" isn't it? When the data gets accessed, if we detect that is stale, we refresh it. It doesn't replace the concept of pruning, which is instead the process of removing the unused and unreferenced entries.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@d5d70762-12d0-45a1-890d-524b12d3f735>